Document Class: PSSMHelper

+++ Code/Native/Source/LightSystem/PSSMHelper.cpp
@@ -1,26 +1,48 @@
 
 #include "PSSMHelper.h"
 
-/*
-
 #include <Eigen/Dense>
 
-
-void generate_equations(Eigen::MatrixXf &eq_system, Eigen::VectorXf &eq_results, const LVector4f &vec, LVector4f expected, int offset) {
-    int write_offset = offset * 4;
-
-    for (int row = 0; row < 4; ++row) {
+/**
+ * @brief Internal method to generate a set of equations
+ * @details This generates a set of equations into a given equation system.
+ *   Based on a given start point and end point, it inserts an equation to the
+ *   system which will, when solved, solve the following equation:
+ *   vec * <SOLVED-TRANSFORM> = expected
+ * 
+ * @param eq_system The equation system
+ * @param eq_results The target result vector of the equation system
+ * @param vec The actual vector to be transformed
+ * @param expected The expected output of the transformation
+ * @param offset The index of the equation
+ */
+void generate_equations(Eigen::MatrixXf &eq_system, Eigen::VectorXf &eq_results, const LVector4f &vec, LVector4f expected, size_t offset) {
+    size_t write_offset = offset * 4;
+    for (size_t row = 0; row < 4; ++row) {
         float expected_coeff = expected.get_cell(row);
-        int col_offset = row * 4;
-        for (int col = 0; col < 4; ++col) {
+        size_t col_offset = row * 4;
+        for (size_t col = 0; col < 4; ++col) {
             eq_system(write_offset, col_offset + col) = vec.get_cell(col);
         }
-
         eq_results(write_offset++) = expected_coeff;
     }
 }
-
-
+/**
+ * @brief Finds a projection mat arround the given set of points.
+ * @details This methods finds a projection matrix which projects the given set
+ *   of frustum points to a unit cube, which can be used as a camera matrix.
+ *   The eight points should determine the frustum uniquely.
+ *   
+ *  @param near_ul The Upper-Left point of the frustum on the near plane
+ *  @param near_ur The Upper-Right point of the frustum on the near plane
+ *  @param near_ll The Lower-Left point of the frustum on the near plane
+ *  @param near_lr The Lower-Right point of the frustum on the near plane
+ *  @param far_ul The Upper-Left point of the frustum on the far plane
+ *  @param far_ur The Upper-Right point of the frustum on the far plane
+ *  @param far_ll The Lower-Left point of the frustum on the far plane
+ *  @param far_lr The Lower-Right point of the frustum on the far plane
+ *   
+ */
 LMatrix4f PSSMHelper::find_projection_mat(
             const LVector4f &near_ul,
             const LVector4f &near_ur,
@@ -40,7 +62,7 @@ LMatrix4f PSSMHelper::find_projection_mat(
     equation_system.fill(0);
 
     // Generate the equations
-    int offset = 0;
+    size_t offset = 0;
     generate_equations(equation_system, equation_results, near_ul, LVector4f(-1,  1, 0, 1), offset++);
     generate_equations(equation_system, equation_results, near_ur, LVector4f( 1,  1, 0, 1), offset++);
     generate_equations(equation_system, equation_results, near_ll, LVector4f(-1, -1, 0, 1), offset++);
@@ -51,16 +73,10 @@ LMatrix4f PSSMHelper::find_projection_mat(
     generate_equations(equation_system, equation_results, far_ll,  LVector4f(-1, -1, 1, 1), offset++);
     generate_equations(equation_system, equation_results, far_lr,  LVector4f( 1, -1, 1, 1), offset++);
 
-
-    // cout << "Equation system: " << endl << equation_system << endl;
-    // cout << "Coeff vector: " << endl << equation_results << endl;
-
-    // Solve equation system
+    // Solve the equation system
     Eigen::VectorXf solved_system = equation_system.colPivHouseholderQr().solve(equation_results);
 
-    // cout << "Solved system: " << endl << solved_system << endl;
-
-    // Construct result matrix and return it
+    // Construct result matrix and return it. We also need to transpose the matrix.
     LMatrix4f result(
             solved_system(0), solved_system(4), solved_system(8), solved_system(12), 
             solved_system(1), solved_system(5), solved_system(9), solved_system(13), 
@@ -69,6 +85,3 @@ LMatrix4f PSSMHelper::find_projection_mat(
         );
     return result;
 }
-
-
-*/
